Fixing orientation changes handling
The ViewManager no longer returns a PhotoView directly. Instead it
returns a PhotoViewWrapper, which has `onSizeChanged` method overriden.
This way when the view changes sizes, it redraws correctly.

`setSource` code has been moved to this wrapper, so that it can be
easily reused in `onSizeChanged`. A boolean `mInitialized` is flagged to
avoid recreating certain structures or setting the image URi, which
would make the redraw flicker and be slower.